+++ back/prisma/migrations/20231213165437_remove_email/migration.sql
@@ -1,0 +1,11 @@
+/*
+  Warnings:
+
+  - You are about to drop the column `email` on the `user` table. All the data in the column will be lost.
+
+*/
+-- DropIndex
+DROP INDEX "user_email_key";
+
+-- AlterTable
+ALTER TABLE "user" DROP COLUMN "email";