+++ back/prisma/migrations/20231213212129_remove_dni/migration.sql
@@ -1,0 +1,11 @@
+/*
+  Warnings:
+
+  - You are about to drop the column `dni` on the `user` table. All the data in the column will be lost.
+
+*/
+-- DropIndex
+DROP INDEX "user_dni_key";
+
+-- AlterTable
+ALTER TABLE "user" DROP COLUMN "dni";