+++ back/src/__tests__/resources/me.test.ts
@@ -1,20 +1,21 @@
 import supertest from 'supertest'
 import { expect, it, describe, beforeAll, afterAll } from 'vitest'
-import { Category } from '@prisma/client'
+import { Category, User } from '@prisma/client'
 import { server, testUserData } from '../globalSetup'
-import { authToken } from '../setup'
 import { pathRoot } from '../../routes/routes'
 import { prisma } from '../../prisma/client'
 import { resourceGetSchema } from '../../schemas'
 import { resourceTestData } from '../mocks/resources'
 import { checkInvalidToken } from '../helpers/checkInvalidToken'
+import { authToken } from '../mocks/ssoServer'
 
+let user: User | null
 beforeAll(async () => {
   const testCategory = (await prisma.category.findUnique({
     where: { slug: 'testing' },
   })) as Category
-  const user = await prisma.user.findUnique({
-    where: { email: testUserData.user.email },
+  user = await prisma.user.findFirst({
+    where: { name: testUserData.user.name },
   })
 
   const testResourcesWithUser = resourceTestData.map((resource) => {
@@ -59,11 +60,11 @@ beforeAll(async () => {
 
 afterAll(async () => {
   await prisma.favorites.deleteMany({
-    where: { user: { email: testUserData.user.email } },
+    where: { user: { id: user?.id } },
   })
   await prisma.vote.deleteMany({})
   await prisma.resource.deleteMany({
-    where: { user: { email: testUserData.user.email } },
+    where: { user: { id: user?.id } },
   })
 })
 
@@ -80,7 +81,7 @@ describe('Testing resources/me endpoint', () => {
     // User admin has no posted resources
     const response = await supertest(server)
       .get(`${pathRoot.v1.resources}/me`)
-      .set('Cookie', authToken.admin)
+      .set('Cookie', [`authToken=${authToken.admin}`])
 
     expect(response.status).toBe(200)
     expect(response.body).toBeInstanceOf(Array)
@@ -91,7 +92,7 @@ describe('Testing resources/me endpoint', () => {
     // Normal user has a resource created for this test.
     const response = await supertest(server)
       .get(`${pathRoot.v1.resources}/me`)
-      .set('Cookie', authToken.user)
+      .set('Cookie', [`authToken=${authToken.user}`])
 
     expect(response.status).toBe(200)
     expect(response.body).toBeInstanceOf(Array)
@@ -105,7 +106,7 @@ describe('Testing resources/me endpoint', () => {
     const testCategorySlug = 'testing'
     const response = await supertest(server)
       .get(`${pathRoot.v1.resources}/me`)
-      .set('Cookie', authToken.user)
+      .set('Cookie', [`authToken=${authToken.user}`])
       .query({ testCategorySlug })
     expect(response.status).toBe(200)
     expect(response.body).toBeInstanceOf(Array)
@@ -116,12 +117,9 @@ describe('Testing resources/me endpoint', () => {
   })
 
   it('If the user voted and favorited one of its own created resources, it should be reflected on the response object', async () => {
-    const user = await prisma.user.findUnique({
-      where: { email: testUserData.user.email },
-    })
     const response = await supertest(server)
       .get(`${pathRoot.v1.resources}/me`)
-      .set('Cookie', authToken.user)
+      .set('Cookie', [`authToken=${authToken.user}`])
     expect(response.body).toEqual(
       expect.arrayContaining([
         expect.objectContaining({