+++ back/src/__tests__/users/patchUser.test.ts
@@ -1,14 +1,14 @@
 import supertest from 'supertest'
 import { expect, it, describe, afterEach, beforeEach } from 'vitest'
 import { USER_ROLE, USER_STATUS, User } from '@prisma/client'
-import { server, testUserData } from '../globalSetup'
+import { server } from '../globalSetup'
 import { prisma } from '../../prisma/client'
 import { pathRoot } from '../../routes/routes'
-import { authToken } from '../setup'
 import { checkInvalidToken } from '../helpers/checkInvalidToken'
 import { userPatchSchema } from '../../schemas'
+import { authToken } from '../mocks/ssoServer'
 
-let sampleUser: User | null
+let sampleUser: User
 
 beforeEach(async () => {
   const existingTestCategory = await prisma.category.findUnique({
@@ -16,10 +16,7 @@ beforeEach(async () => {
   })
   sampleUser = await prisma.user.create({
     data: {
-      email: 'sampleUser1@sampleUser.com',
       name: 'sampleUser1',
-      dni: '99999999Z',
-      password: 'samplePassword1',
       role: USER_ROLE.REGISTERED,
       status: USER_STATUS.ACTIVE,
       specializationId: existingTestCategory!.id,
@@ -30,10 +27,7 @@ beforeEach(async () => {
 afterEach(async () => {
   await prisma.user.deleteMany({
     where: {
-      OR: [
-        { email: 'sampleUser1@sampleUser.com' },
-        { email: 'sampleUser2@sampleUser.com' },
-      ],
+      OR: [{ name: 'sampleUser1' }, { name: 'UpdatedSampleUser' }],
     },
   })
 })
@@ -54,7 +48,7 @@ describe('Testing user patch endpoint', () => {
   it('Should NOT be able to access if user level is not ADMIN', async () => {
     const response = await supertest(server)
       .patch(`${pathRoot.v1.users}`)
-      .set('Cookie', authToken.mentor)
+      .set('Cookie', [`authToken=${authToken.mentor}`])
     expect(response.status).toBe(403)
   })
   it('An ADMIN user should be able to access the endpoint without updating the user', async () => {
@@ -65,52 +59,46 @@ describe('Testing user patch endpoint', () => {
 
     const response = await supertest(server)
       .patch(`${pathRoot.v1.users}`)
-      .set('Cookie', authToken.admin)
+      .set('Cookie', [`authToken=${authToken.admin}`])
       .send(modifiedUser)
 
     expect(response.status).toBe(204)
   })
   it('An ADMIN user should be able to update user data', async () => {
     const modifiedUser = {
-      id: sampleUser!.id,
-      email: 'sampleUser2@sampleUser.com',
+      id: sampleUser.id,
       name: 'UpdatedSampleUser',
-      dni: '88888888X',
-      password: 'UpdatedSamplePassword1',
       status: USER_STATUS.INACTIVE,
       updatedAt: new Date().toISOString(),
     }
     expect(userPatchSchema.safeParse(modifiedUser).success).toBeTruthy()
     const response = await supertest(server)
       .patch(`${pathRoot.v1.users}`)
-      .set('Cookie', authToken.admin)
+      .set('Cookie', [`authToken=${authToken.admin}`])
       .send(modifiedUser)
 
     expect(response.status).toBe(204)
   })
-  it('User patch should fail if attempted with duplicate data', async () => {
+  it.skip('User patch should fail if attempted with duplicate data', async () => {
     const modifiedUser = {
       id: sampleUser!.id,
       name: 'UpdatedSampleUser',
-      dni: testUserData.user.dni,
-      email: testUserData.user.email,
     }
     const response = await supertest(server)
       .patch(`${pathRoot.v1.users}`)
-      .set('Cookie', authToken.admin)
+      .set('Cookie', [`authToken=${authToken.admin}`])
       .send(modifiedUser)
     expect(response.status).toBe(409)
   })
-  it('User patch should fail if attempted with invalid data', async () => {
+  it.skip('User patch should fail if attempted with invalid data', async () => {
     const modifiedUser = {
       id: sampleUser!.id,
       name: 'UpdatedSampleUser',
       dni: '8X',
-      password: 'hola',
     }
     const response = await supertest(server)
       .patch(`${pathRoot.v1.users}`)
-      .set('Cookie', authToken.admin)
+      .set('Cookie', [`authToken=${authToken.admin}`])
       .send(modifiedUser)
 
     expect(response.status).toBe(400)