+++ back/src/schemas/users/userRegisterSchema.ts
@@ -2,7 +2,7 @@ import { userSchema } from './userSchema'
 import { z } from '../../openapi/zod'
 import { passwordRegex } from '../regex/passwordRegex'
 
-export const userRegisterSchema = userSchema
+export const userRegSchema = userSchema
   .pick({
     email: true,
     dni: true,
@@ -13,10 +13,14 @@ export const userRegisterSchema = userSchema
     accept: z.literal<boolean>(true),
     password: z.string().min(8).regex(passwordRegex),
     confirmPassword: z.string().min(8).regex(passwordRegex),
+    itineraryId: z.string().nonempty().cuid(),
   })
-  .refine((data) => data.password === data.confirmPassword, {
+export const userRegisterSchema = userRegSchema.refine(
+  (data) => data.password === data.confirmPassword,
+  {
     message: 'Passwords must match',
     path: ['confirmPassword'],
-  })
+  }
+)
 
 export type UserRegister = z.infer<typeof userRegisterSchema>