+++ packages/ui/src/components/atoms/Checkbox.tsx
@@ -12,7 +12,6 @@ export type TCheckbox = InputHTMLAttributes<HTMLInputElement> & {
   className?: string
   onChange?: (e: React.ChangeEvent<HTMLInputElement>) => void
   disabled?: boolean
-  legalTerms?: boolean
 }
 
 export const CheckBoxWrapper = styled(FlexBox)`
@@ -26,24 +25,10 @@ export const CheckBoxWrapper = styled(FlexBox)`
     border-radius: ${dimensions.borderRadius.xs};
     border: 2px solid ${colors.gray.gray3};
     cursor: pointer;
-    // margin-right: ${dimensions.spacing.xxs};
-    accent-color: ${colors.primaryLight};
-
-    /* &:hover,
-    &:focus {
-      background-color: ${colors.primaryLight};
-      border: 2px solid ${colors.primary};
-    } */
 
     &:checked {
       background-color: ${colors.primary};
       border: 2px solid ${colors.primary};
-    }
-
-    &:checked + :hover,
-    :focus {
-      background-color: ${colors.primaryDark};
-      border: 2px solid ${colors.primaryDark};
     }
 
     &:checked::before {
@@ -58,6 +43,7 @@ export const CheckBoxWrapper = styled(FlexBox)`
 
     &:disabled {
       border: 2px solid ${colors.gray.gray4};
+      cursor: not-allowed;
     }
   }
 
@@ -65,56 +51,18 @@ export const CheckBoxWrapper = styled(FlexBox)`
     font-weight: 500;
     color: ${colors.gray.gray2};
     cursor: pointer;
-    &:hover,
-    &:focus {
-      font-weight: 600;
-      color: ${colors.gray.gray1};
-    }
   }
 
   > input:checked + label {
     font-weight: 600;
     color: ${colors.gray.gray1};
-
-    &:hover,
-    &:focus {
-      font-weight: 500;
-      color: ${colors.gray.gray2};
-    }
   }
 
   > input:disabled + label {
     color: ${colors.gray.gray4};
+    cursor: not-allowed;
   }
-
-  ${({ legalTerms }) =>
-    legalTerms &&
-    `margin-top: ${dimensions.spacing.base};
-  margin-bottom: ${dimensions.spacing.base};
-
-  > input {
-    border: 1px solid ${colors.black.black1};
-
-    &:checked {
-      border: 2px solid ${colors.primary};
-    }
-  }
-  `}
 `
-
-//ESTILS NEGRE FORM: const CheckBoxStyled = styled(CheckBox)`
-//   margin-top: ${dimensions.spacing.base};
-//   margin-bottom: ${dimensions.spacing.base};
-
-//   > input {
-//     border: 1px solid ${colors.black.black1};
-
-//     &:checked {
-//       border: 2px solid ${colors.primary};
-//     }
-//   }
-// `
-// //FAltA^ra : coMPORTMAENT CHECKED, ONcHANGE??
 
 const Checkbox = React.forwardRef<HTMLInputElement, TCheckbox>(
   (
@@ -127,18 +75,17 @@ const Checkbox = React.forwardRef<HTMLInputElement, TCheckbox>(
       className,
       onChange,
       disabled = false,
-      legalTerms = false,
       ...rest
     },
     ref
   ) => (
     <CheckBoxWrapper
       className={className}
-      justify='flex-start'
+      justify="flex-start"
       gap={`${dimensions.spacing.xxxs}`}
     >
       <input
-        type='checkbox'
+        type="checkbox"
         ref={ref}
         id={id}
         name={id}
@@ -146,7 +93,6 @@ const Checkbox = React.forwardRef<HTMLInputElement, TCheckbox>(
         required={required}
         onChange={onChange}
         disabled={disabled}
-        legalTerms={legalTerms}
         {...rest}
       />
       <Label htmlFor={id} text={label} hidden={hiddenLabel} />