+++ packages/ui/src/stories/atoms/Checkbox.stories.ts
@@ -1,4 +1,3 @@
-import React from 'react'
 import type { Meta, StoryObj } from '@storybook/react'
 import Checkbox from '../../components/atoms/Checkbox'
 
@@ -8,7 +7,7 @@ const meta = {
   component: Checkbox,
   parameters: {
     // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/configure/story-layout
-    layout: 'centered'
+    layout: 'centered',
   },
   // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/writing-docs/autodocs
   tags: ['autodocs'],
@@ -22,8 +21,7 @@ const meta = {
     defaultChecked: { control: 'boolean', defaultValue: false },
     required: { control: 'boolean', defaultValue: false },
     disabled: { control: 'boolean', defaultValue: false },
-    legalTerms: { control: 'boolean', defaultValue: false }
-  }
+  },
 } satisfies Meta<typeof Checkbox>
 
 export default meta
@@ -33,22 +31,22 @@ type Story = StoryObj<typeof meta>
 export const Default: Story = {
   args: {
     id: 'checkboxId',
-    label: 'A checkbox'
-  }
+    label: 'A checkbox',
+  },
 }
 
 export const DefaultChecked: Story = {
   args: {
     id: 'checkedCheckboxId',
     label: 'A checked by default checkbox',
-    defaultChecked: true
-  }
+    defaultChecked: true,
+  },
 }
 
 export const Disabled: Story = {
   args: {
     id: 'disabledCheckboxId',
     label: 'A disabled checkbox',
-    disabled: true
-  }
+    disabled: true,
+  },
 }