+++ sso/src/routes/tokenRoutes.ts
@@ -4,6 +4,7 @@ import { validate as validateToken } from '../controllers/token/validate'
 import { validate } from '../middleware/validate'
 import { z } from '../openapi/zod'
 import { validateSchema } from '../schemas/token/validateSchema'
+import { authenticate } from '../middleware/authenticate'
 
 export const tokenRoutes = new Router()
 
@@ -12,5 +13,6 @@ tokenRoutes.prefix(pathRoot.v1.tokens)
 tokenRoutes.post(
   '/validate',
   validate(z.object({ body: validateSchema })),
+  authenticate,
   validateToken
 )